Better structure for VIsualPath and Layer
Polyline construction, and tracking of layers and extrusions/travels is
done.

Tube geometry generation has not been reincorporated yet, so the
rendering is just done with lines.

Stubs and pseudocode are in place for a few central functions that will
need to be written: updateVisibleTubeRanges,
updateVisiblePolylineRanges, setVisibleLayerRange, and
setVisibleCommandRange.

The gcode index and layerindex sliders do not work yet, but once the
above functions are implemented it will be trivial to fix.